requirements.txt
Click==7.0
Flask==1.0.2
gevent==1.4.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
python-dotenv==0.10.2
Werkzeug==0.15.4
pymongo==3.10.1
